backend/requirements.txt
# SmartSense Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Date/Time
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
